old-tests/submission/W3C/domhtml/HTMLBaseElement01.html

<!DOCTYPE html>
<html><head>
<base href="http://天気の良い日.w3c-test.org/">
<meta charset="utf-8">
<title>NIST DOM HTML Test - Base</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-base-element.html#dom-base-href" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
    The href attribute specifies the base URI. 
</p>
<p>
    Retrieve the href attribute and examine its value.  
</p>


<script>
var domtest = async_test("The href attribute specifies the base URI.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vhref;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("base");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vhref = testNode.href;

      assert_equals(vhref,"http://天気の良い日.w3c-test.org/");
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
